requirements.txt
Flask
cryptography
psycopg2-binary
flask-sqlalchemy
python-dotenv